Reducers extracted, repos downloading fine.
