+++ app/src/main/java/com/example/zipporeppogithub/ui/repos/ReposViewModel.kt
@@ -14,7 +14,6 @@ import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import okhttp3.ResponseBody
@@ -89,8 +88,11 @@ class ReposViewModel
             isAllGranted = isAllGranted && isGranted
         }
         if (isAllGranted) {
+            //Возможно ConcurrentModificationException, если подобный вызов не блокирует содержимое reposToDownload
+            val localReposToDownload = buildList { addAll(reposToDownload) }
+            reposToDownload.clear()
             viewModelScope.launch(Dispatchers.IO) {
-                reposToDownload.forEach { loadZipRepo(it) } //todo ConcurrentModificationException, что вполне логично
+                localReposToDownload.forEach { loadZipRepo(it) }
             }
         } else {
             reposToDownload.clear()
@@ -196,67 +198,6 @@ class ReposViewModel
                             reducer.sendEvent(ReposEvent.SetError(errMsg))
                         }
                     }
-                }
-            }
-        }
-    }
-
-    private class ReposReducer(initial: ReposState) {
-
-        private val _state: MutableStateFlow<ReposState> = MutableStateFlow(initial)
-        val state: StateFlow<ReposState>
-            get() = _state
-
-        fun setState(newState: ReposState) {
-            _state.tryEmit(newState)
-        }
-
-        fun sendEvent(event: ReposEvent) {
-            reduce(_state.value, event)
-        }
-
-        fun reduce(oldState: ReposState, event: ReposEvent) {
-            when (event) {
-                is ReposEvent.NavigateToHtml -> {
-                    setState(oldState.copy(htmlLink = event.html))
-                }
-                is ReposEvent.ScreenNavigateOut -> {
-                    setState(oldState.copy(htmlLink = null))
-                }
-                is ReposEvent.PermissionRequested -> {
-                    setState(
-                        oldState.copy(
-                            isPermissionRequired = event.isRequested
-                        )
-                    )
-                }
-                is ReposEvent.ReposLoading -> {
-                    setState(
-                        oldState.copy(
-                            isLoading = true,
-                            errorMsg = null
-                        )
-                    )
-                }
-                is ReposEvent.NewReposFound -> {
-                    val newReposList = oldState.repos.toMutableList()
-                    newReposList.addAll(event.repos)
-                    setState(
-                        oldState.copy(
-                            repos = newReposList,
-                            errorMsg = null,
-                            isLoading = false
-                        )
-                    )
-                }
-                is ReposEvent.NoReposFound -> {
-                    //todo
-                }
-                is ReposEvent.SetError -> {
-                    setState(oldState.copy(errorMsg = event.errorMsgResource, isLoading = false))
-                }
-                is ReposEvent.ShowSnack -> {
-                    //todo
                 }
             }
         }
